btis/backend/requirements.txt
flask==3.0.0
flask-sqlalchemy==3.1.1
flask-jwt-extended==4.6.0
flask-cors==4.0.0
flask-mail==0.9.1
flask-socketio==5.3.6
werkzeug==3.0.1
numpy==1.26.2
pandas==2.1.4
scikit-learn==1.3.2
joblib==1.3.2
python-dotenv==1.0.0
bcrypt==4.1.2
marshmallow==3.20.1
requests==2.31.0
eventlet==0.33.3
apscheduler==3.10.4
